P-Lambda constraint is enforced by default. This helped to avoid very small dT in transition from Picard to BDF1 in one test case out of one.
Cleaned the use of flag is_matrix_symmetric. It is no longer used
to calculate relative permeability.

Implementaiton of a new nonlinear solver has been commenced. refs #922

+++ src/flow/Richards_AztecOO.cpp
@@ -57,7 +57,7 @@ void Richards_PK::SolveFullySaturatedProblem(double Tp, Epetra_Vector& u)
   solver_tmp->SetRHS(&b);
 
   solver_tmp->SetLHS(&u);
-  solver_tmp->Iterate(max_itrs_linear, convergence_tol_linear);
+  solver_tmp->Iterate((long long)max_itrs_linear, convergence_tol_linear);
 
   if (MyPID == 0 && verbosity >= FLOW_VERBOSITY_HIGH) {
     int num_itrs = solver_tmp->NumIters();
@@ -107,7 +107,7 @@ void Richards_PK::SolveTransientProblem(double Tp, double dTp, Epetra_Vector& u)
   solver_tmp->SetRHS(&b);
 
   solver_tmp->SetLHS(&u);
-  solver_tmp->Iterate(max_itrs_linear, convergence_tol_linear);
+  solver_tmp->Iterate((long long)max_itrs_linear, convergence_tol_linear);
 
   if (MyPID == 0 && verbosity >= FLOW_VERBOSITY_HIGH) {
     int num_itrs = solver_tmp->NumIters();
@@ -150,7 +150,7 @@ void Richards_PK::EnforceConstraints_MFD(double Tp, Epetra_Vector& u)
   solver_tmp->SetRHS(&b);
 
   solver_tmp->SetLHS(&utmp);
-  solver_tmp->Iterate(max_itrs_linear, convergence_tol_linear);
+  solver_tmp->Iterate((long long)max_itrs_linear, convergence_tol_linear);
   *u_faces = *utmp_faces;
 
   if (MyPID == 0 && verbosity >= FLOW_VERBOSITY_HIGH) {
